+++ src/types/field.ts
@@ -1,0 +1,30 @@
+/**
+ * NullableString
+ */
+export type NullableString = string | null;
+
+/**
+ * Result
+ */
+export interface FieldResult {
+  /**
+   * Status
+   *
+   * @type {boolean}
+   */
+  ok: boolean;
+
+  /**
+   * Value
+   *
+   * @type {string | number | boolean | null}
+   */
+  value: string | number | boolean | null;
+
+  /**
+   * Error
+   *
+   * @type {string}
+   */
+  error?: string;
+}